+++ frontend/src/components/common/index.ts
@@ -2,10 +2,11 @@ import Avatar from './Avatar';
 import Button from './Button';
 import CubeIcon from './CubeIcon';
 import Dropdown from './Dropdown';
+import HeadMeta from './HeadMeta';
 import LanguageIcon from './LanguageIcon';
 import Paper from './Paper';
 import Searchbar from './Searchbar';
 import { RankingSkeleton } from './Skeleton';
 import Spinner from './Spinner';
 
-export { LanguageIcon, Button, Dropdown, Avatar, Searchbar, Spinner, CubeIcon, RankingSkeleton, Paper };
+export { LanguageIcon, Button, Dropdown, Avatar, Searchbar, Spinner, CubeIcon, RankingSkeleton, Paper, HeadMeta };
