+++ frontend/src/pages/ranking.tsx
@@ -7,12 +7,12 @@ import { QueryClient, dehydrate, useQuery } from '@tanstack/react-query';
 import { CubeRankType } from '@type/common';
 import { RankingPaiginationResponse } from '@type/response';
 import Filterbar from '@components/Filterbar';
-import HeadMeta from '@components/HeadMeta';
 import Pagination from '@components/Pagination';
 import RankingTable from '@components/Ranking';
 import NotFound from '@components/Ranking/NotFound';
 import RankingSearchbar from '@components/Ranking/RankingSearchbar';
 import { Avatar, CubeIcon, LanguageIcon, RankingSkeleton } from '@components/common';
+import HeadMeta from '@components/common/HeadMeta';
 import { requestTokenRefresh } from '@apis/auth';
 import { requestTopRankingByScore } from '@apis/ranking';
 import { COUNT_PER_PAGE, CUBE_RANK } from '@utils/constants';
